refactor: 提取 PerfPage 的 P/Invoke 到独立 NativeMethods.cs (PerfPage 2616→2399 行)
